virt-handler: Move VSOCK code outside of virt-handler and extract method
Moved the package "pkg/virt-handler/vsock" to "pkg/vsock".
In the next commit, the new package will contain common VSOCK code
also used by virt-launcher.

Extracted the code block that creates VSOCK connection
into a separate method.
