+++ cmd/virt-handler/BUILD.bazel
@@ -45,7 +45,7 @@ go_library(
         "//pkg/virt-handler/rest:go_default_library",
         "//pkg/virt-handler/seccomp:go_default_library",
         "//pkg/virt-handler/selinux:go_default_library",
-        "//pkg/virt-handler/vsock:go_default_library",
+        "//pkg/vsock/server:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
