+++ cmd/virt-handler/virt-handler.go
@@ -88,7 +88,7 @@ import (
 	"kubevirt.io/kubevirt/pkg/virt-handler/rest"
 	"kubevirt.io/kubevirt/pkg/virt-handler/seccomp"
 	"kubevirt.io/kubevirt/pkg/virt-handler/selinux"
-	"kubevirt.io/kubevirt/pkg/virt-handler/vsock"
+	"kubevirt.io/kubevirt/pkg/vsock/server"
 )
 
 const (
@@ -305,7 +305,7 @@ func (app *virtHandlerApp) Run() {
 		logger.Criticalf("Error constructing migration tls config: %v", err)
 		os.Exit(2)
 	}
-	vsockMgr := vsock.NewVSOCKHypervisorService(1, app.caManager)
+	vsockMgr := server.NewVSOCKHypervisorService(1, app.caManager)
 
 	vsockConfigCallback := func() {
 		if app.clusterConfig.VSOCKEnabled() {
