+++ hack/linter/lint-paths.txt
@@ -61,7 +61,6 @@ pkg/virtctl/version
 pkg/virt-handler/cgroup/constants
 pkg/virt-handler/ksm
 pkg/virt-handler/node-labeller/util
-pkg/virt-handler/vsock
 pkg/virt-launcher/metadata
 pkg/virt-launcher/premigration-hook-server/network
 pkg/virt-launcher/virtwrap/access-credentials
@@ -73,6 +72,7 @@ pkg/virt-launcher/virtwrap/converter/virtio
 pkg/virt-launcher/virtwrap/network
 pkg/virt-launcher/virtwrap/statsconv/util
 pkg/vmitrait
+pkg/vsock
 tests/conformance
 tests/console
 tests/decorators
