+++ pkg/vsock/server/BUILD.bazel
@@ -3,11 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["server.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-handler/vsock",
+    importpath = "kubevirt.io/kubevirt/pkg/vsock/server",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/util/tls:go_default_library",
-        "//pkg/virt-handler/vsock/system:go_default_library",
+        "//pkg/vsock/system:go_default_library",
         "//pkg/vsock/system/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/mdlayher/vsock:go_default_library",
